\name{jags.View}
% functions
\alias{jags.View}

\title{View a jagsUI output object in a separate window}

\usage{

 jags.View(x, title, digits=3)

}

\arguments{
  \item{x}{A jagsUI object}
  \item{title}{Specify a title for the window.}
  \item{digits}{Number of digits to display after the decimal.}
}

\description{
Show an R object in a separate, spreadsheet-style window via a call to \code{\link[utils]{View}}.
}

\author{
  Ken Kellner \email{contact@kenkellner.com} and Mike Meredith.
}
