\name{traceplots}
% functions
\alias{traceplots}

\title{Traceplots of JAGS output}

\usage{
  traceplots(x)
}

\arguments{
  \item{x}{A simplejags object}
}

\description{
  Displays a series of MCMC iteration plots for each monitored parameter in a JAGS analysis.
}

\author{ 
  Ken Kellner \email{ken@kenkellner.com}.
}